Separated out utility functions from `contact.js` and added tests
