+++ addon/content/stub.xhtml
@@ -21,6 +21,7 @@
   <script type="text/javascript" src="vendor/redux-toolkit.umd.js"></script>
   <script type="text/javascript" src="vendor/reactjs-popup.js"></script>
   <script type="text/javascript" src="vendor/prop-types.js"></script>
+  <script type="text/javascript" src="utils.js"></script>
   <script type="text/javascript" src="reducer.js"></script>
   <script type="text/javascript" src="svgIcon.js"></script>
   <script type="text/javascript" src="attachments.js"></script>
